+++ src/gql/documents/queries/preference/preferenceEventLocationsAll.ts
@@ -4,7 +4,7 @@ import type { AllPreferenceEventLocationsQueryVariables } from '~~/gql/generated
 
 export const allPreferenceEventLocationsQuery = graphql(`
   query AllPreferenceEventLocations {
-    allAccountPreferenceEventLocations {
+    allPreferenceEventLocations {
       nodes {
         ...PreferenceEventLocationItem
       }